#88 - Constructed outline of Gotcha Level. Organize file hierarchy a bit
BP_IKGameInstance, BP_IKGameModeBase, BP_IKGotchaGameMode, BP_IKHUD, BP_IKPlayerController - Organize file hierarchy.
BP_GotchaWidget, BP_IKGotchaGameMode, BP_IKGotchaHUD, GotchaLevel - Constructed Outline of GotchaLevel.
CombatResultUI, ItemPickerUI, IKGameModeBase, IKPlayerController - Reinclude organized header files.
