+++ Config/DefaultEngine.ini
@@ -2,7 +2,7 @@
 
 [/Script/EngineSettings.GameMapsSettings]
 GameDefaultMap=/Engine/Maps/Templates/OpenWorld
-GameInstanceClass=/Game/__BluePrints/BP_IKGameInstance.BP_IKGameInstance_C
+GameInstanceClass=/Game/__BluePrints/WorldSettings/BP_IKGameInstance.BP_IKGameInstance_C
 
 [/Script/WindowsTargetPlatform.WindowsTargetSettings]
 DefaultGraphicsRHI=DefaultGraphicsRHI_DX12
