+++ Source/Project_IK/Private/WorldSettings/IKGameModeBase.cpp
@@ -13,13 +13,12 @@ See LICENSE file in the project root for full license information.
 
 #include "Kismet/GameplayStatics.h"
 
-#include "Project_IK/Public/UI/IKHUD.h"
 #include "WorldSettings/IKGameInstance.h"
 #include "WorldSettings/IKPlayerController.h"
 #include "Components/TargetingComponent.h"
 #include "Managers/LevelTransitionManager.h"
 
-#include "UI/IKHUD.h"
+#include "WorldSettings/IKHUD.h"
 
 
 #include "Characters/HeroBase.h"
