+++ Source/Project_IK/Public/UI/GotchaWidget.h
@@ -1,0 +1,17 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Blueprint/UserWidget.h"
+#include "GotchaWidget.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class PROJECT_IK_API UGotchaWidget : public UUserWidget
+{
+	GENERATED_BODY()
+	
+};
