+++ Source/Project_IK/Public/WorldSettings/GotchaWorld/IKGotchaGameMode.h
@@ -1,0 +1,25 @@
+/******************************************************************************
+Copyright(C) 2024
+Author: sinil.kang(rtd99062@gmail.com)
+Creation Date : 12.28.2024
+Summary : Header file of game mode for Gotcha level.
+
+Licensed under the MIT License.
+See LICENSE file in the project root for full license information.
+******************************************************************************/
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/GameModeBase.h"
+#include "IKGotchaGameMode.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class PROJECT_IK_API AIKGotchaGameMode : public AGameModeBase
+{
+	GENERATED_BODY()
+	
+};
